+++ lib/src/config.dart
@@ -1,0 +1,11 @@
+part of ace;
+
+abstract class Config extends Disposable {
+  
+  /// Returns the module url for the given [modulePath] and [moduleType].
+  String moduleUrl(String modulePath, String moduleType);
+  
+  /// Sets the module url for the given [modulePath] to the given [moduleUrl]
+  /// and returns the [moduleUrl].
+  String setModuleUrl(String modulePath, String moduleUrl);  
+}