+++ lib/src/implementation.dart
@@ -56,4 +56,6 @@ abstract class Implementation {
   }
   
   require(String modulePath);
+  
+  Config get config;
 }