+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.5.1-dev+7.1.2014
 
 - Updated to ace-builds version 07.01.2014.
+- Added the `Config` abstract class and the `ace.config` top-level getter.
 
 ## 0.5.0+6.9.2014